+++ public/data/expiredAccessToken.json
@@ -1,0 +1,4 @@
+{
+  "errorCode": 403,
+  "errorMsg": "ACCESS토큰 만료"
+}
